+++ changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Added** [#60](https://github.com/codyburleson/cburleson-pwa/issues/60) Page: base22-consultants-creed
 - **Added** [#67](https://github.com/codyburleson/cburleson-pwa/issues/67) Page: eclipse-tip-format-source-code-on-save
 - **Added** [#61](https://github.com/codyburleson/cburleson-pwa/issues/61) Page: better-error-messages-from-gulp-using-gulp-util
+- **Added** [#62](https://github.com/codyburleson/cburleson-pwa/issues/62) Page: binding-to-a-zsphere-armature
 
 ## [1.1.3] - 2019-12-26